requirements.txt
discord.py
flask
certifi